+++ srcs/asm_fct.c
@@ -6,34 +6,11 @@
 /*   By: mdos-san <mdos-san@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/11 18:44:45 by mdos-san          #+#    #+#             */
-/*   Updated: 2017/03/03 11:42:45 by mdos-san         ###   ########.fr       */
+/*   Updated: 2017/03/03 14:38:04 by mdos-san         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "corewar.h"
-
-int		mod(int a, int b)
-{
-	int	i;
-
-	i = a + b;
-	i %= MEM_SIZE;
-	if (i < 0)
-		i = MEM_SIZE + i;
-	return (i);
-}
-
-int		idx_mod(int a)
-{
-	int	n;
-
-	n = (a < 0) ? 1 : 0;
-	a *= (a < 0) ? -1 : 1;
-	a %= IDX_MOD;
-	if (n)
-		a *= -1;
-	return (a);
-}
 
 void	assign_live(t_cw *cw, int number, int champ_offset)
 {
@@ -56,132 +33,6 @@ void	live(t_cw *cw, t_process *p)
 	p->nb_live += 1;
 }
 
-void	ld(t_cw *cw, t_process *p)
-{
-	int		y;
-	int		value;
-	t_ocp	ocp;
-
-	y = 0;
-	cw->nb_param = 2;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	ocp_parse(cw, p, ocp, 0, 1);
-	value = *p->p_one;
-	if (ft_strcmp(ocp.one, "11") == 0)
-	{
-		*p->p_one = (int)(short)(*p->p_one);
-		*p->p_one = idx_mod(*p->p_one);
-		board_to_int(cw, &value, *p->p_one);
-	}
-	if (cw->param_error == 0)
-	{
-		while (y < 16)
-		{
-			if (p->r + y == p->p_two)
-				*p->p_two = value;
-			++y;
-		}
-		p->carry = (*p->p_two == 0) ? 1 : 0;
-	}
-}
-
-void	st(t_cw *cw, t_process *p)
-{
-	t_ocp	ocp;
-
-	cw->nb_param = 2;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	ocp_parse(cw, p, ocp, 0, 1);
-	if (cw->param_error == 0)
-	{
-		if (ft_strcmp(ocp.two, "01") == 0)
-			*p->p_two = *p->p_one;
-		else
-		{
-			*p->p_two = (int)(short)(*p->p_two);
-			*p->p_two = idx_mod(*p->p_two);
-			int_to_board(cw, p->p_one, *p->p_two);
-		}
-	}
-}
-
-void	add(t_cw *cw, t_process *p)
-{
-	int		tmp;
-	t_ocp	ocp;
-
-	cw->nb_param = 3;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	ocp_parse(cw, p, ocp, 1, 0);
-	tmp = *p->p_one + *p->p_two;
-	*p->p_three = tmp;
-	if (*p->p_three == 0)
-		p->carry = 1;
-	else
-		p->carry = 0;
-}
-
-void	sub(t_cw *cw, t_process *p)
-{
-	int		tmp;
-	t_ocp	ocp;
-
-	cw->nb_param = 3;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	ocp_parse(cw, p, ocp, 1, 0);
-	tmp = *p->p_one - *p->p_two;
-	*p->p_three = tmp;
-	if (*p->p_three == 0)
-		p->carry = 1;
-	else
-		p->carry = 0;
-}
-
-void	and(t_cw *cw, t_process *p)
-{
-	t_ocp	ocp;
-
-	cw->nb_param = 3;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	cw->idx = 1;
-	ocp_parse(cw, p, ocp, 0, 0);
-	*p->p_three = *p->p_one & *p->p_two;
-	if (*p->p_three == 0)
-		p->carry = 1;
-	else
-		p->carry = 0;
-}
-
-void	or(t_cw *cw, t_process *p)
-{
-	t_ocp	ocp;
-
-	cw->nb_param = 3;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	cw->idx = 1;
-	ocp_parse(cw, p, ocp, 0, 0);
-	*p->p_three = *p->p_one | *p->p_two;
-	if (*p->p_three == 0)
-		p->carry = 1;
-	else
-		p->carry = 0;
-}
-
-void	xor(t_cw *cw, t_process *p)
-{
-	t_ocp	ocp;
-
-	cw->nb_param = 3;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	cw->idx = 1;
-	ocp_parse(cw, p, ocp, 0, 0);
-	*p->p_three = *p->p_one ^ *p->p_two;
-	if (*p->p_three == 0)
-		p->carry = 1;
-	else
-		p->carry = 0;
-}
-
 void	zjmp(t_cw *cw, t_process *p)
 {
 	short	j;
@@ -196,117 +47,6 @@ void	zjmp(t_cw *cw, t_process *p)
 	}
 	else
 		cw->nb_readed = 3;
-}
-
-void	ldi(t_cw *cw, t_process *p)
-{
-	int		new_addr;
-	t_ocp	ocp;
-
-	cw->nb_param = 3;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	ocp_parse(cw, p, ocp, 1, 0);
-	new_addr = mod(*p->p_one, *p->p_two);
-	board_to_int(cw, p->p_three, new_addr);
-}
-
-void	frk(t_cw *cw, t_process *p)
-{
-	short		j;
-	t_process	*child;
-	int			i;
-
-	i = 0;
-	j = 0;
-	((unsigned char *)&j)[1] = cw->board[mod(p->pc, 1)];
-	((unsigned char *)&j)[0] = cw->board[mod(p->pc, 2)];
-	j = idx_mod(j);
-	child = process_new(&cw->process, p->nb_champ,
-			mod(p->pc, j), p->color_nb);
-	while (i < 16)
-	{
-		child->r[i] = p->r[i];
-		++i;
-	}
-	child->carry = p->carry;
-	child->nb_live = p->nb_live;
-	cw->nb_readed = 3;
-}
-
-void	sti(t_cw *cw, t_process *p)
-{
-	int		sum;
-	t_ocp	ocp;
-
-	cw->nb_param = 3;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	cw->idx = 1;
-	ocp_parse(cw, p, ocp, 1, 0);
-	if (ft_strcmp(ocp.two, "10") == 0)
-		*p->p_two = (int)(short)(*p->p_two);
-	if (ft_strcmp(ocp.three, "10") == 0)
-		*p->p_three = (int)(short)(*p->p_three);
-	sum = *p->p_two + *p->p_three;
-	sum = idx_mod(sum);
-	int_to_board(cw, p->p_one, sum);
-}
-
-void	lld(t_cw *cw, t_process *p)
-{
-	int		y;
-	t_ocp	ocp;
-
-	y = 0;
-	cw->nb_param = 2;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	ocp_parse(cw, p, ocp, 0, 0);
-	while (y < 16)
-	{
-		if (p->r + y == p->p_two)
-		{
-			*p->p_two = *p->p_one;
-			break ;
-		}
-		++y;
-	}
-	if (*p->p_two == 0)
-		p->carry = 1;
-	else
-		p->carry = 0;
-}
-
-void	lldi(t_cw *cw, t_process *p)
-{
-	int		new_addr;
-	t_ocp	ocp;
-
-	cw->nb_param = 3;
-	ocp = ocp_get(cw->board[mod(p->pc, 1)]);
-	ocp_parse(cw, p, ocp, 1, 0);
-	new_addr = mod(*p->p_one, *p->p_two);
-	board_to_int(cw, p->p_three, new_addr);
-}
-
-void	lfrk(t_cw *cw, t_process *p)
-{
-	unsigned short	j;
-	t_process		*child;
-	int				i;
-
-	i = 0;
-	j = 0;
-	((unsigned char *)&j)[1] = cw->board[mod(p->pc, 1)];
-	((unsigned char *)&j)[0] = cw->board[mod(p->pc, 2)];
-	j = (int)(short)(j);
-	child = process_new(&cw->process, p->nb_champ, mod(p->pc, j), p->color_nb);
-	while (i < 16)
-	{
-		child->r[i] = p->r[i];
-		++i;
-	}
-	child->carry = p->carry;
-	child->nb_live = 1;
-	cw->nb_readed = 3;
 }
 
 void	aff(t_cw *cw, t_process *p)
